For #1506: Show resume in /join if user already has resume. - added puzzle for creating Resumes.resume(login) - Extracted /join-post logic into TkJoinPost and left /join logic into TkJoin - Fixed TkApp according new TkJoin and TkJoinPost classes - Added resume.xsl - Added todo for implementation.
