+++ src/main/java/com/zerocracy/tk/TkApp.java
@@ -163,14 +163,8 @@ public final class TkApp extends TkWrap {
                                                                     new FkRegex("/spam-send", new TkSpam(farm)),
                                                                     new FkRegex("/shutdown", new TkShutdown(props, farm)),
                                                                     new FkRegex("/policy", new TkPolicy()),
-                                                                    new FkRegex(
-                                                                        "/join",
-                                                                        (Take) req -> {
-                                                                            new RqUser(farm, req, false).value();
-                                                                            return new RsPage(farm, "/xsl/join.xsl", req);
-                                                                        }
-                                                                    ),
-                                                                    new FkRegex("/join-post", new TkJoin(farm)),
+                                                                    new FkRegex("/join", new TkJoin(farm)),
+                                                                    new FkRegex("/join-post", new TkJoinPost(farm)),
                                                                     new FkRegex(
                                                                         "/org/takes/.+\\.xsl",
                                                                         new TkClasspath()
